Refactor existing testing utils into a helpers package
