+++ timetable_solutions/tests/integration/domain/data_management/upload_processors/test_timetable_slot.py
@@ -18,7 +18,8 @@ from data import constants as data_constants
 from data import models
 from domain.data_management import upload_processors
 from domain.data_management.constants import Header
-from tests import data_factories, utils
+from tests import data_factories
+from tests.helpers.csv import get_csv_from_lists
 
 
 @pytest.mark.django_db
@@ -31,7 +32,7 @@ class TestTimetableSlotFileUploadProcessor:
         yg_2 = data_factories.YearGroup(school=school)
 
         # Get a csv file-like object
-        csv_file = utils.get_csv_from_lists(
+        csv_file = get_csv_from_lists(
             [
                 [
                     Header.SLOT_ID,
@@ -104,7 +105,7 @@ class TestTimetableSlotFileUploadProcessor:
         invalid_row[missing_column_index] = None
 
         # Get a csv file-like object
-        csv_file = utils.get_csv_from_lists(
+        csv_file = get_csv_from_lists(
             [
                 [
                     Header.SLOT_ID,
@@ -137,7 +138,7 @@ class TestTimetableSlotFileUploadProcessor:
         school = data_factories.School()
 
         # Get a csv file-like object
-        csv_file = utils.get_csv_from_lists(
+        csv_file = get_csv_from_lists(
             [
                 [
                     Header.SLOT_ID,
